+++ src/testserial.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python3
+
+# Test some ideas for the scanner
+
+from scanner import Scanner
+from time import sleep
+
+s = Scanner()
+
+print("We have our scanner:", s)
+
+print("Writing a simple GLG command")
+s.writeline('GLG')
+print("Retrieving the rsponse")
+print(s.readline())
+
+print("Now send a few commands with decoded response")
+
+print("MDL:", s.command("MDL"))
+print("VER:", s.command("VER"))
+print("GLG:", s.command("GLG"))
+
+# Run a loop for a minute (more or less)
+print("Looping for 5 receptions. Use ^C if it's taking too long.")
+t = 0
+try:
+    while t < 5:
+        r = s.command("GLG")
+        if r.NAME1:
+            print("Sys={NAME1}, Group={NAME2}, Chan={NAME3}, Sql={SQL}, Mute={MUT}, Time={TIME}".format_map(r))
+            t += 1
+        sleep(1)
+except KeyboardInterrupt:
+    print("\nOK, quitting the loop")
+finally:
+    print("Loop done...")