+++ CHANGELOG.md
@@ -1,6 +1,21 @@
-## CHANGELOG
+### CHANGELOG
 
 *Top down (newest first) list of major changes*
+
+####15 Jul 2014
+Added 'vol' command and associated commands and options.
+
+####6 Jul 2014
+Scanner operational and initial test OK
+
+Formatters started. GLG and STS supplied.
+
+*ToDo:*
+* Change Response class to add vars to the namespace instead of the var dict
+
+* Add missing attribute handler to return ?None? (something else?) instead of throwing an error.
+
+* Perhaps only for all UPPERCASE attributes?
 
 ####26 June 2014